refactor main.py to minimal content
creates the following:

 * mycli/cli_runner.py
 * mycli/client.py
 * mycli/client_commands.py
 * mycli/client_connection.py
 * mycli/client_query.py

and reduces main.py to only:

 * click_entrypoint()
 * main()
