+++ changelog.md
@@ -1,3 +1,11 @@
+Upcoming (TBD)
+==============
+
+Internal
+---------
+* Factor `main.py` into several files using mixins.
+
+
 1.73.1 (2026/05/29)
 ==============
