+++ test/pytests/test_main_modes_batch.py
@@ -12,6 +12,7 @@ from typing import Any, Literal, cast
 from click.testing import CliRunner
 import pytest
 
+import mycli.cli_runner as cli_runner
 import mycli.main_modes.batch as batch_mode
 import test.pytests.test_main as test_main_module
 import test.utils as test_utils
@@ -120,6 +121,7 @@ def patch_progress_mode(monkeypatch, mycli_main, mycli_main_batch) -> None:
     monkeypatch.setattr(mycli_main_batch, 'ProgressBar', DummyProgressBar)
     monkeypatch.setattr(mycli_main_batch.prompt_toolkit.output, 'create_output', lambda **kwargs: object())
     fake_sys = make_fake_sys(stdin_tty=False, stderr_tty=True)
+    monkeypatch.setattr(cli_runner, 'sys', fake_sys)
     monkeypatch.setattr(mycli_main, 'sys', fake_sys)
     monkeypatch.setattr(mycli_main_batch, 'sys', fake_sys)
 
@@ -341,7 +343,9 @@ def test_main_batch_with_progress_bar_rejects_non_files(monkeypatch, tmp_path) -
     result = main_batch_with_progress_bar(DummyMyCli(), cli_args)
 
     assert result == 1
-    assert messages == [('--progress is only compatible with a plain file.', True, 'red')]
+    assert '--progress is only compatible with a plain file.' in messages[0][0]
+    assert messages[0][1] is True
+    assert messages[0][2] == 'red'
 
 
 def test_main_batch_with_progress_bar_handles_open_errors(monkeypatch) -> None:
